+++ app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Category;
 use App\Models\SubCategory;
 use App\Models\Product;
+use Illuminate\Support\Facades\Validator;
 
 
 class ProductController extends Controller
@@ -18,13 +19,22 @@ class ProductController extends Controller
     }
     public function productInsert(Request $request)
     {
-        $request->validate([
+        $validateRequest = Validator::make($request->all(), [
             'name' => 'required',
             'category_id' => 'required',
             'subcategory_id' => 'required',
+            'description' => 'required',
             'price' => 'required',
             'quantity' => 'required',
         ]);
+
+        if ($validateRequest->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Validation error',
+                'errors' => $validateRequest->errors()
+            ], 403);
+        }
 
         $product = Product::create([
             'name' => $request->input('name'),
@@ -36,11 +46,18 @@ class ProductController extends Controller
         ]);
 
         session()->flash('success', 'Product added successfully!');
+        return response()->json(['message' => 'User registered successfully', 'product' => $product], 201);
+
         return redirect()->route('products');
     }
     public function products()
     {
         $products = Product::with('category', 'subcategory')->get();
+        return response()->json([
+            'success' => true,
+            'message' => 'Product Data successfully',
+            'result' => $products
+        ], 200);
         return view('products.view_product', compact('products'));
     }
     public function productEdit($id)
@@ -53,7 +70,7 @@ class ProductController extends Controller
     }
     public function productUpdate(Request $request, $id)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'required',
             'category_id' => 'required',
             'subcategory_id' => 'required',
@@ -61,7 +78,19 @@ class ProductController extends Controller
             'quantity' => 'required',
         ]);
 
+        if ($validator->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Validation fails',
+                'error' => $validator->errors()
+            ], 401);
+        }
+
         $products = Product::find($id);
+
+        if (is_null($products)) {
+            return response()->json(['message' => 'User not found'], 404);
+        }
 
         $products->update([
             'name' => $request->input('name'),
@@ -73,12 +102,20 @@ class ProductController extends Controller
         ]);
 
         session()->flash('success', 'Product Update successfully!');
+        return response()->json([
+            'message' => 'Product updated successfully',
+            'user' => $products,
+        ], 200);
         return redirect()->route('products');
     }
     public function productDestroy($id)
     {
         $product = Product::find($id);
+        if (!$product) {
+            return response()->json(['message' => 'User not found'], 404);
+        }
         $product->delete();
+        return response()->json(['message' => 'Product deleted successfully']);
         session()->flash('danger', 'Product Delete successfully!');
         return redirect()->back();
     }
