+++ app/Http/Controllers/SubCategoryController.php
@@ -10,6 +10,13 @@ class SubCategoryController extends Controller
 {
     public function subcategory(){
         $subcategorys = SubCategory::with('category')->get();
+
+        return response()->json([
+            'success' => true,
+            'message' => 'Sub Category Data successfully',
+            'result' => $subcategorys
+        ], 200);
+
         return view('subcategory.view_sub_category',compact('subcategorys'));
     }
 
@@ -31,7 +38,7 @@ class SubCategoryController extends Controller
             $image->move('images', $filename);
         }
         
-        SubCategory::create([
+        $subcategory = SubCategory::create([
             'category_id'      => $request->input('category_id'),
             'subcategory_name' => $request->input('subcategory_name'),
             'image'            => $filename,
@@ -39,6 +46,13 @@ class SubCategoryController extends Controller
         ]);
 
         session()->flash('success', 'Subcategory added successfully!');
+
+        return response()->json([
+            'success' => true,
+            'message' => 'Sub Category inserted successfully.',
+            'data' => $subcategory
+        ], 201);
+
         return redirect()->route('subcategory');
     }
     public function Editsubcategory($id)
@@ -56,6 +70,11 @@ class SubCategoryController extends Controller
         ]);
     
         $subcategory = SubCategory::find($id);
+
+        if (!$subcategory) {
+            return response()->json(['error' => 'Sub Category not found'], 404);
+        }
+
         if ($request->hasFile('image')) {
             $image = $request->file('image');
             $filename = time() . '.' . $image->getClientOriginalExtension();
@@ -66,14 +85,18 @@ class SubCategoryController extends Controller
             'category_id' => $request->input('category_id'),
             'subcategory_name' => $request->input('subcategory_name'),
         ]);
-
+        return response()->json(['success' => 'Sub Category updated successfully.', 'category' => $subcategory], 200);
         session()->flash('success', 'Subcategory Updated successfully!');
         return redirect()->route('subcategory');
     }
     public function Destroysubcategory($id)
     {
         $subcategory = SubCategory::find($id);
+        if (!$subcategory) {
+            return response()->json(['message' => 'Sub Category not found'], 404);
+        }
         $subcategory->delete();
+        return response()->json(['message' => 'Sub Category deleted successfully']);
         session()->flash('danger', 'Subcategory Delete successfully!');
         return redirect()->back();
     }
