+++ app/Http/Controllers/UserController.php
@@ -5,12 +5,20 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
 {
     public function users()
     {
         $users = User::all();
+
+        return response()->json([
+            'success' => true,
+            'message' => 'Users Data successfully',
+            'result' => $users
+        ], 200);
         return view('user.view_user', compact('users'));
     }
     public function userCreate()
@@ -20,21 +28,33 @@ class UserController extends Controller
 
     public function userInsert(Request $request)
     {
-        $request->validate([
+        $validateRequest = Validator::make($request->all(), [
             'name' => 'required',
             'email' => 'required',
             'address' => 'required',
             'mobile_no' => 'required',
+            'password' => 'required',
         ]);
+
+        if ($validateRequest->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Validation error',
+                'errors' => $validateRequest->errors()
+            ], 403);
+        }
+
 
         $user = User::create([
             'name'      => $request->input('name'),
             'email'     => $request->input('email'),
             'address'  => $request->input('address'),
             'mobile_no'   => $request->input('mobile_no'),
+            'password' => Hash::make($request->password),
         ]);
 
         session()->flash('success', 'User added successfully!');
+        return response()->json(['message' => 'User registered successfully', 'user' => $user], 201);
         return redirect()->route('user');
     }
 
@@ -46,31 +66,52 @@ class UserController extends Controller
 
     public function userUpdate(Request $request, $id)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'required',
             'email' => 'required',
             'address' => 'required',
             'mobile_no' => 'required',
         ]);
 
+        if ($validator->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Validation fails',
+                'error' => $validator->errors()
+            ], 401);
+        }
+
         $users = User::find($id);
+
+        if (is_null($users)) {
+            return response()->json(['message' => 'User not found'], 404);
+        }
 
         $users->update([
             'name'      => $request->input('name'),
             'email'     => $request->input('email'),
             'address'  => $request->input('address'),
             'mobile_no'   => $request->input('mobile_no'),
+            'password' => Hash::make($request->password),
         ]);
 
         session()->flash('success', 'User Update successfully!');
+        return response()->json([
+            'message' => 'User updated successfully',
+            'user' => $users,
+        ], 200);
         return redirect()->route('user');
     }
 
     public function userDestroy($id)
     {
         $user = User::find($id);
+        if (!$user) {
+            return response()->json(['message' => 'User not found'], 404);
+        }
         $user->delete();
         session()->flash('danger', 'User Delete successfully!');
+        return response()->json(['message' => 'User deleted successfully']);
         return redirect()->back();
     }
 
@@ -95,12 +136,20 @@ class UserController extends Controller
     public function Profileupdate(Request $request, $id)
     {
 
-        $request->validate([
+        $validateRequest = Validator::make($request->all(), [
             'name' => 'required',
             'email' => 'required',
             'address' => 'required',
             'mobile_no' => 'required',
         ]);
+
+        if ($validateRequest->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Validation error',
+                'errors' => $validateRequest->errors()
+            ], 403);
+        }
 
         $users = User::find($id);
 
@@ -112,6 +161,7 @@ class UserController extends Controller
         ]);
 
         return redirect()->route('myprofile')->with('success', 'Profile updated successfully');
+        
     }
     
 }
